rename / use CamelCase and move to subnamespace => 1 folder = 1 namespace
